requirements.txt
flask
pillow
qrcode[pil]
numpy
opencv-python
